+++ lib/ui/widgets/expandable_text.dart
@@ -33,6 +33,8 @@ class ExpandableText extends StatefulWidget {
   final String? readMoreIconAsset;
   final EdgeInsetsGeometry readMoreIconPadding;
 
+  final Widget? footerWidget;
+
   const ExpandableText(this.text, {
     Key? key,
     this.textStyle,
@@ -48,6 +50,8 @@ class ExpandableText extends StatefulWidget {
     this.readMoreIcon,
     this.readMoreIconAsset,
     this.readMoreIconPadding = const EdgeInsets.only(left: 7),
+
+    this.footerWidget,
   })  : super(key: key);
 
   TextStyle get _textStyle => textStyle ?? TextStyle(fontFamily: Styles().fontFamilies?.regular, fontSize: 16, color: Styles().colors?.textBackground,);
@@ -74,7 +78,9 @@ class _ExpandableTextState extends State<ExpandableText> {
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
-
+      
+      List<Widget> contentList = <Widget>[];
+      
       TextPainter textPainter = TextPainter(
         textScaleFactor: MediaQuery.of(context).textScaleFactor,
         textDirection: TextDirection.rtl,
@@ -94,11 +100,9 @@ class _ExpandableTextState extends State<ExpandableText> {
       if (textPainter.didExceedMaxLines) {
         
         String displayText = _collapsed ? widget.text.substring(0, endIndex) + widget.trimSuffix : widget.text;
-        List<Widget> contentList = <Widget>[
-          RichText(textScaleFactor: MediaQuery.of(context).textScaleFactor, softWrap: true, overflow: TextOverflow.clip,
-            text: TextSpan(text: displayText, style: widget._textStyle,),
-          ),
-        ];
+        contentList.add(RichText(textScaleFactor: MediaQuery.of(context).textScaleFactor, softWrap: true, overflow: TextOverflow.clip,
+          text: TextSpan(text: displayText, style: widget._textStyle,),
+        ),);
 
         if (_collapsed) {
           List<Widget> readMoreList = <Widget>[
@@ -121,12 +125,19 @@ class _ExpandableTextState extends State<ExpandableText> {
             ),
           ]);
         }
-        
-        return Column(children: contentList);
+        else if (widget.footerWidget != null) {
+          contentList.add(widget.footerWidget!);
+        }
       }
       else {
-        return Text(widget.text, style: widget._textStyle,);
+        contentList.add(Text(widget.text, style: widget._textStyle,));
+
+        if (widget.footerWidget != null) {
+          contentList.add(widget.footerWidget!);
+        }
       }
+
+      return Column(crossAxisAlignment: CrossAxisAlignment.start, children: contentList);
     },);
   }