+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added footer widget in ExpandableText [#3055](https://github.com/rokwire/illinois-app/issues/3055).
 
 ## [1.3.2] - 2023-02-16
 ### Changed